requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.9
django-crispy-forms==1.6.1
django-formtools==1.0
django-pure-pagination==0.3.0
django-reversion==2.0.8
django-simple-captcha==0.4.6
DjangoUeditor==1.8.143
future==0.15.2
get==0.0.20
httplib2==0.9.2
kombu==4.0.2
MySQL-python==1.2.5
olefile==0.44
Pillow==3.4.2
post==0.0.12
public==0.0.37
PyMySQL==0.7.10
pytz==2017.2
query-string==0.0.11
setupfiles==0.0.49
six==1.10.0
vine==1.1.3
virtualenv==15.1.0
virtualenvwrapper-win==1.2.1
XlsxWriter==0.9.6
xlwt==1.2.0
